+++ axelor-crm/src/main/java/com/axelor/apps/crm/service/OpportunityServiceImpl.java
@@ -25,6 +25,7 @@ import com.axelor.apps.crm.db.Lead;
 import com.axelor.apps.crm.db.Opportunity;
 import com.axelor.apps.crm.db.repo.OpportunityRepository;
 import com.axelor.apps.crm.exception.IExceptionMessage;
+import com.axelor.apps.message.db.EmailAddress;
 import com.axelor.exception.AxelorException;
 import com.axelor.exception.db.repo.TraceBackRepository;
 import com.axelor.i18n.I18n;
@@ -69,6 +70,11 @@ public class OpportunityServiceImpl implements OpportunityService {
       address.setFullName(addressService.computeFullName(address));
     }
 
+    EmailAddress email = null;
+    if (lead.getEmailAddress() != null) {
+      email = new EmailAddress(lead.getEmailAddress().getAddress());
+    }
+
     Partner partner =
         Beans.get(PartnerService.class)
             .createPartner(
@@ -76,7 +82,7 @@ public class OpportunityServiceImpl implements OpportunityService {
                 null,
                 lead.getFixedPhone(),
                 lead.getMobilePhone(),
-                lead.getEmailAddress(),
+                email,
                 opportunity.getCurrency(),
                 address,
                 address);